이 부분도 누락되어있네요 ㅠ
주석 부분은 누락 된거고 dbmanager에 추가된 권한검사 함수 추가 된 부분 수정

+++ src/main/java/OOSE/controller/MemberManagement/DeleteMember.java
@@ -14,17 +14,15 @@ import java.io.PrintWriter;
 public class DeleteMember extends HttpServlet
 {
     private MemberDBManager dbManager = new MemberDBManager();
-    int authority=3;
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException
     {
-        //로그인 구현 쪽 병합시 테스트해볼 예정
-//        if (checkAuthority((int) req.getSession().getAttribute("authority")))
-//        {
-//            printAlert("권한이 없습니다",resp);
-//            return;
-//        }
+        if(!checkAuthority((int) req.getSession().getAttribute("authority")))
+        {
+            printAlert("권한이 없습니다",resp);
+            return;
+        }
         Member member = new Member();
         member.setId(req.getParameter("id"));
         if(!dbManager.deleteMemberInfo(member)) //삭제 실패한 경우
@@ -36,7 +34,7 @@ public class DeleteMember extends HttpServlet
     }
     public boolean checkAuthority(int authority)
     {
-        if(authority>this.authority||authority==1)      //회원이거나 관리자일 경우
+        if(authority>dbManager.findAuthority("회원삭제")||authority==1)      //회원이거나 관리자일 경우
             return true;
         return false;       //권한 없음
     }
